requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
openai==1.3.0
duckduckgo-search==3.9.6
newspaper3k==0.2.8
dateparser==1.2.0
pydantic==2.5.0
dashscope==1.14.1
streamlit==1.38.0
